# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface ContactInfo {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  github?: string | null;
}

export interface WorkExperience {
  id: string;
  company: string;
  title: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  current?: boolean;
  /** @nullable */
  location?: string | null;
  bullets?: string[];
}

export interface Education {
  id: string;
  institution: string;
  degree: string;
  /** @nullable */
  field?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  current?: boolean;
  /** @nullable */
  gpa?: string | null;
}

export interface Project {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  url?: string | null;
  technologies?: string[];
  bullets?: string[];
}

export interface Certification {
  id: string;
  name: string;
  /** @nullable */
  issuer?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  url?: string | null;
}

export interface Achievement {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
}

export interface Language {
  id: string;
  name: string;
  level: string;
}

export interface CustomSectionItem {
  id: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  description?: string | null;
  bullets?: string[];
}

export interface CustomSection {
  id: string;
  heading: string;
  items?: CustomSectionItem[];
}

export interface Resume {
  id: number;
  title: string;
  template?: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  contact?: ContactInfo;
  workExperience?: WorkExperience[];
  education?: Education[];
  skills?: string[];
  projects?: Project[];
  certifications?: Certification[];
  achievements?: Achievement[];
  languages?: Language[];
  customSections?: CustomSection[];
  createdAt: string;
  updatedAt: string;
}

export interface ResumeInput {
  title: string;
  template?: string;
}

export interface ResumeUpdate {
  title?: string;
  template?: string;
  fullName?: string;
  jobTitle?: string;
  summary?: string;
  photoUrl?: string;
  contact?: ContactInfo;
  workExperience?: WorkExperience[];
  education?: Education[];
  skills?: string[];
  projects?: Project[];
  certifications?: Certification[];
  achievements?: Achievement[];
  languages?: Language[];
  customSections?: CustomSection[];
}

export interface GenerateSummaryRequest {
  fullName: string;
  jobTitle: string;
  workExperience?: WorkExperience[];
  skills?: string[];
}

export interface ImproveBulletRequest {
  bullet: string;
  jobTitle: string;
  /** @nullable */
  company?: string | null;
}

export interface SuggestSkillsRequest {
  jobTitle: string;
  existingSkills?: string[];
}

export type ScoreResumeRequestResumeData = { [key: string]: unknown };

export interface ScoreResumeRequest {
  resumeData: ScoreResumeRequestResumeData;
}

export interface ScoreFeedbackItem {
  category: string;
  score: number;
  maxScore: number;
  feedback: string;
  tips?: string[];
}

export interface ResumeScoreResult {
  totalScore: number;
  maxScore: number;
  grade: string;
  summary?: string;
  breakdown: ScoreFeedbackItem[];
}

export interface GeneratedTextResult {
  text: string;
}

export interface SuggestedSkillsResult {
  skills: string[];
}