# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CV Builder API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: resumes
    description: Resume CRUD operations
  - name: ai
    description: AI-powered content generation

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List all resumes
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: createResume
      tags: [resumes]
      summary: Create a new resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Created resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a resume by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateResume
      tags: [resumes]
      summary: Update a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeUpdate"
      responses:
        "200":
          description: Updated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /resumes/{id}/duplicate:
    post:
      operationId: duplicateResume
      tags: [resumes]
      summary: Duplicate a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /ai/generate-summary:
    post:
      operationId: generateSummary
      tags: [ai]
      summary: Generate a professional summary using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSummaryRequest"
      responses:
        "200":
          description: Generated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedTextResult"

  /ai/improve-bullet:
    post:
      operationId: improveBullet
      tags: [ai]
      summary: Improve a bullet point using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImproveBulletRequest"
      responses:
        "200":
          description: Improved bullet point
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedTextResult"

  /ai/suggest-skills:
    post:
      operationId: suggestSkills
      tags: [ai]
      summary: Suggest skills based on job title and experience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestSkillsRequest"
      responses:
        "200":
          description: Suggested skills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestedSkillsResult"

  /ai/score-resume:
    post:
      operationId: scoreResume
      tags: [ai]
      summary: Score a resume and provide improvement tips
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreResumeRequest"
      responses:
        "200":
          description: Resume score and feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeScoreResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ContactInfo:
      type: object
      properties:
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        linkedin:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        github:
          type: ["string", "null"]

    WorkExperience:
      type: object
      required: [id, company, title, startDate]
      properties:
        id:
          type: string
        company:
          type: string
        title:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        current:
          type: boolean
        location:
          type: ["string", "null"]
        bullets:
          type: array
          items:
            type: string

    Education:
      type: object
      required: [id, institution, degree, startDate]
      properties:
        id:
          type: string
        institution:
          type: string
        degree:
          type: string
        field:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        current:
          type: boolean
        gpa:
          type: ["string", "null"]

    Project:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        technologies:
          type: array
          items:
            type: string
        bullets:
          type: array
          items:
            type: string

    Certification:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        issuer:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        url:
          type: ["string", "null"]

    Achievement:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]

    Language:
      type: object
      required: [id, name, level]
      properties:
        id:
          type: string
        name:
          type: string
        level:
          type: string

    CustomSectionItem:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        bullets:
          type: array
          items:
            type: string

    CustomSection:
      type: object
      required: [id, heading]
      properties:
        id:
          type: string
        heading:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CustomSectionItem"

    Resume:
      type: object
      required: [id, title, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        template:
          type: string
        fullName:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        contact:
          $ref: "#/components/schemas/ContactInfo"
        workExperience:
          type: array
          items:
            $ref: "#/components/schemas/WorkExperience"
        education:
          type: array
          items:
            $ref: "#/components/schemas/Education"
        skills:
          type: array
          items:
            type: string
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        certifications:
          type: array
          items:
            $ref: "#/components/schemas/Certification"
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
        languages:
          type: array
          items:
            $ref: "#/components/schemas/Language"
        customSections:
          type: array
          items:
            $ref: "#/components/schemas/CustomSection"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ResumeInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        template:
          type: string

    ResumeUpdate:
      type: object
      properties:
        title:
          type: string
        template:
          type: string
        fullName:
          type: string
        jobTitle:
          type: string
        summary:
          type: string
        photoUrl:
          type: string
        contact:
          $ref: "#/components/schemas/ContactInfo"
        workExperience:
          type: array
          items:
            $ref: "#/components/schemas/WorkExperience"
        education:
          type: array
          items:
            $ref: "#/components/schemas/Education"
        skills:
          type: array
          items:
            type: string
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        certifications:
          type: array
          items:
            $ref: "#/components/schemas/Certification"
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
        languages:
          type: array
          items:
            $ref: "#/components/schemas/Language"
        customSections:
          type: array
          items:
            $ref: "#/components/schemas/CustomSection"

    GenerateSummaryRequest:
      type: object
      required: [fullName, jobTitle]
      properties:
        fullName:
          type: string
        jobTitle:
          type: string
        workExperience:
          type: array
          items:
            $ref: "#/components/schemas/WorkExperience"
        skills:
          type: array
          items:
            type: string

    ImproveBulletRequest:
      type: object
      required: [bullet, jobTitle]
      properties:
        bullet:
          type: string
        jobTitle:
          type: string
        company:
          type: ["string", "null"]

    SuggestSkillsRequest:
      type: object
      required: [jobTitle]
      properties:
        jobTitle:
          type: string
        existingSkills:
          type: array
          items:
            type: string

    ScoreResumeRequest:
      type: object
      required: [resumeData]
      properties:
        resumeData:
          type: object

    ScoreFeedbackItem:
      type: object
      required: [category, score, maxScore, feedback]
      properties:
        category:
          type: string
        score:
          type: integer
        maxScore:
          type: integer
        feedback:
          type: string
        tips:
          type: array
          items:
            type: string

    ResumeScoreResult:
      type: object
      required: [totalScore, maxScore, grade, breakdown]
      properties:
        totalScore:
          type: integer
        maxScore:
          type: integer
        grade:
          type: string
        summary:
          type: string
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/ScoreFeedbackItem"

    GeneratedTextResult:
      type: object
      required: [text]
      properties:
        text:
          type: string

    SuggestedSkillsResult:
      type: object
      required: [skills]
      properties:
        skills:
          type: array
          items:
            type: string