# lib/api-zod/src/generated/types/certification.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Certification {
  id: string;
  name: string;
  /** @nullable */
  issuer?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  url?: string | null;
}