# lib/api-zod/src/generated/types/contactInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInfo {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  github?: string | null;
}