# lib/api-zod/src/generated/types/customSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomSectionItem } from './customSectionItem';

export interface CustomSection {
  id: string;
  heading: string;
  items?: CustomSectionItem[];
}