# lib/api-zod/src/generated/types/customSectionItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomSectionItem {
  id: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  description?: string | null;
  bullets?: string[];
}