# lib/api-zod/src/generated/types/resume.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Achievement } from './achievement';
import type { Certification } from './certification';
import type { ContactInfo } from './contactInfo';
import type { CustomSection } from './customSection';
import type { Education } from './education';
import type { Language } from './language';
import type { Project } from './project';
import type { WorkExperience } from './workExperience';

export interface Resume {
  id: number;
  title: string;
  template?: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  contact?: ContactInfo;
  workExperience?: WorkExperience[];
  education?: Education[];
  skills?: string[];
  projects?: Project[];
  certifications?: Certification[];
  achievements?: Achievement[];
  languages?: Language[];
  customSections?: CustomSection[];
  createdAt: Date;
  updatedAt: Date;
}