# lib/api-zod/src/generated/types/resumeScoreResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreFeedbackItem } from './scoreFeedbackItem';

export interface ResumeScoreResult {
  totalScore: number;
  maxScore: number;
  grade: string;
  summary?: string;
  breakdown: ScoreFeedbackItem[];
}