# lib/api-zod/src/generated/types/scoreFeedbackItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreFeedbackItem {
  category: string;
  score: number;
  maxScore: number;
  feedback: string;
  tips?: string[];
}